S10P22A405-298-남상엽-전적검색-컨트롤러-연동

+++ src/main/java/com/ssafy/matchup_statistics/indicator/entity/match/MatchIndicatorStatistics.java
@@ -45,7 +45,7 @@ public class MatchIndicatorStatistics {
         private int isFinishedBeforeFifteenCount;
         private int isOurTeamEarlySurrenderedCount;
         private int winCount;
-        private int pingCountAvg;
+        private double pingCountAvg;
         private double killAvg;
         private double deathAvg;
         private double assistAvg;
